Fix scope spaghetti in init()
Only do the configuration assignation in success() callback for
Ajax getConfiguration call, and do the window/extjs widgets creation
in the regular init() flow.

Allows us to remove that ugly initThis variable, and to properly
use this.options when needed, this way manifest.json correctly
set default options values, and config.json overrides them.

As a bonus, it seems to fix the addition of the 'cadastrapp' item
to the tools menu, which still has a strange behaviour but will
be fixed later on.