cadastrapp - dockerization
This commit introduces a new maven profile which aimes to make the
docker image generation easier, following the conventions from the
geOrchestra project.

It will require a datadir which is not migrated yet, see:
https://github.com/pmauduit/config/tree/cadastrapp-docker

Note: the introduced dockerfile unpacks the cadastrapp webapp, so that
it is easier to inject a jetty-env.xml file, for JNDI configuration.

Tested (compilation only) with the following command:

```
mvn clean package docker:build -Pdocker                                \
-DdockerDatadirScmUrl=scm:git:https://github.com/pmauduit/config.git   \
-DdockerDatadirScmVersion=cadastrapp-docker
```