docker - fix dockerfile
To reflect some evolutions on the official docker Jetty image, we need
to escalate as root user to uncompress the webapp and copy the jetty-env
file.

Long story short: to be able to handle arm architecture, upstream jetty
image had to drop a native library which was in charge to downgrade as
jetty user when launching the servlet container. This caused the use of
the "USER jetty" directive in the upstream dockerfile, but makes our
dockerfile failing.

Also updated to a newer version of the image

Tests: tested on the C2C private CI, should not harm at runtime.