Improve resilency of the jsbuild script.
It now force the use of python2 and ensure virtualenv is using python2.
This script could potentialy fail since new version of distribution like
debian, red hat, etc... now use python3 as the default version in the
system. Also remove the dep install in CI since the script takes care of
it. This script is based on the one used in mapfishapp in georchestra.